fixes #117

+++ BlOrders2023/Views/FillOrdersPage.xaml.cs
@@ -21,6 +21,7 @@ using BlOrders2023.Reporting;
 using BlOrders2023.Services;
 using System.Drawing.Printing;
 using BlOrders2023.Core.Services;
+using Syncfusion.UI.Xaml.DataGrid;
 
 namespace BlOrders2023.Views;
 
@@ -220,6 +221,25 @@ public sealed partial class FillOrdersPage : Page
         Scanline.Focus(FocusState.Programmatic);
     }
 
+    private async void OrderedItems_RecordDeleting(object sender, Syncfusion.UI.Xaml.DataGrid.RecordDeletingEventArgs e)
+    {
+        //e.Cancel = true;
+        //ContentDialog dialog = new()
+        //{
+        //    XamlRoot = XamlRoot,
+        //    Title = "Confirm Delete",
+        //    Content = "Are you sure you want to delete all shipping items from this order? These changes cannot be undone.",
+        //    PrimaryButtonText = "Delete",
+        //    CloseButtonText = "Cancel",
+        //};
+        //var result = await dialog.ShowAsync();
+        //if (result == ContentDialogResult.Primary)
+        //{
+        //    //TODO DELETE THE ITEM
+        //}
+        
+    }
+
     private async void OrderedItems_RecordDeleted(object sender, Syncfusion.UI.Xaml.DataGrid.RecordDeletedEventArgs e)
     {
         if (e.Items.FirstOrDefault() is ShippingItem)
@@ -247,9 +267,20 @@ public sealed partial class FillOrdersPage : Page
 
     private async void DeleteAll_Click(object sender, RoutedEventArgs e)
     {
-        await ViewModel.DeleteAllShippingItemsAsync();
-        await TrySaveOrderAsync();
-        OrderedVsReceivedGrid.View.Refresh();
+        ContentDialog dialog = new()
+        {
+            XamlRoot = XamlRoot,
+            Title = "Confirm Delete",
+            Content = "Are you sure you want to delete all shipping items from this order? These changes cannot be undone.",
+            PrimaryButtonText = "Delete",
+            CloseButtonText = "Cancel",
+        };
+        var result = await dialog.ShowAsync();
+        if(result == ContentDialogResult.Primary){
+            await ViewModel.DeleteAllShippingItemsAsync();
+            await TrySaveOrderAsync();
+            OrderedVsReceivedGrid.View.Refresh();
+        }
     }
 
     private void PrintOrderFlyoutItem_Click(object sender, RoutedEventArgs e)
